+++ layer-atlas/src/main/java/com/layer/atlas/adapters/AtlasMessagesAdapter.java
@@ -18,6 +18,7 @@ import com.layer.atlas.R;
 import com.layer.atlas.messagetypes.AtlasCellFactory;
 import com.layer.atlas.messagetypes.MessageStyle;
 import com.layer.atlas.util.IdentityRecyclerViewEventListener;
+import com.layer.atlas.util.Log;
 import com.layer.atlas.util.Util;
 import com.layer.sdk.LayerClient;
 import com.layer.sdk.messaging.Identity;
@@ -207,19 +208,19 @@ public class AtlasMessagesAdapter extends RecyclerView.Adapter<AtlasMessagesAdap
      * @return This AtlasMessagesAdapter.
      */
     public AtlasMessagesAdapter addCellFactories(AtlasCellFactory... cellFactories) {
-        for (AtlasCellFactory CellFactory : cellFactories) {
-            CellFactory.setStyle(mMessageStyle);
-            mCellFactories.add(CellFactory);
+        for (AtlasCellFactory cellFactory : cellFactories) {
+            cellFactory.setStyle(mMessageStyle);
+            mCellFactories.add(cellFactory);
 
             mViewTypeCount++;
-            CellType me = new CellType(true, CellFactory);
+            CellType me = new CellType(true, cellFactory);
             mCellTypesByViewType.put(mViewTypeCount, me);
-            mMyViewTypesByCell.put(CellFactory, mViewTypeCount);
+            mMyViewTypesByCell.put(cellFactory, mViewTypeCount);
 
             mViewTypeCount++;
-            CellType notMe = new CellType(false, CellFactory);
+            CellType notMe = new CellType(false, cellFactory);
             mCellTypesByViewType.put(mViewTypeCount, notMe);
-            mTheirViewTypesByCell.put(CellFactory, mViewTypeCount);
+            mTheirViewTypesByCell.put(cellFactory, mViewTypeCount);
         }
         return this;
     }
@@ -382,6 +383,9 @@ public class AtlasMessagesAdapter extends RecyclerView.Adapter<AtlasMessagesAdap
             for (Map.Entry<Identity, Message.RecipientStatus> entry : statuses.entrySet()) {
                 // Only show receipts for other members
                 if (entry.getKey().equals(mLayerClient.getAuthenticatedUser())) continue;
+                // Skip receipts for members no longer in the conversation
+                if (entry.getValue() == null) continue;
+
                 switch (entry.getValue()) {
                     case READ:
                         readCount++;
@@ -535,16 +539,28 @@ public class AtlasMessagesAdapter extends RecyclerView.Adapter<AtlasMessagesAdap
         mFooterPosition = mQueryController.getItemCount();
         updateRecipientStatusPosition();
         notifyDataSetChanged();
+
+        if (Log.isPerfLoggable()) {
+            Log.perf("Messages adapter - onQueryDataSetChanged");
+        }
     }
 
     @Override
     public void onQueryItemChanged(RecyclerViewController controller, int position) {
         notifyItemChanged(position);
+
+        if (Log.isPerfLoggable()) {
+            Log.perf("Messages adapter - onQueryItemChanged. Position: " + position);
+        }
     }
 
     @Override
     public void onQueryItemRangeChanged(RecyclerViewController controller, int positionStart, int itemCount) {
         notifyItemRangeChanged(positionStart, itemCount);
+
+        if (Log.isPerfLoggable()) {
+            Log.perf("Messages adapter - onQueryItemRangeChanged. Position start: " + positionStart + " Count: " + itemCount);
+        }
     }
 
     @Override
@@ -554,6 +570,10 @@ public class AtlasMessagesAdapter extends RecyclerView.Adapter<AtlasMessagesAdap
         notifyItemInserted(position);
         if (mAppendListener != null && (position + 1) == getItemCount()) {
             mAppendListener.onMessageAppend(this, getItem(position));
+        }
+
+        if (Log.isPerfLoggable()) {
+            Log.perf("Messages adapter - onQueryItemInserted. Position: " + position);
         }
     }
 
@@ -566,6 +586,10 @@ public class AtlasMessagesAdapter extends RecyclerView.Adapter<AtlasMessagesAdap
         if (mAppendListener != null && (positionEnd + 1) == getItemCount()) {
             mAppendListener.onMessageAppend(this, getItem(positionEnd));
         }
+
+        if (Log.isPerfLoggable()) {
+            Log.perf("Messages adapter - onQueryItemRangeInserted. Position start: " + positionStart + " Count: " + itemCount);
+        }
     }
 
     @Override
@@ -573,6 +597,10 @@ public class AtlasMessagesAdapter extends RecyclerView.Adapter<AtlasMessagesAdap
         mFooterPosition--;
         updateRecipientStatusPosition();
         notifyItemRemoved(position);
+
+        if (Log.isPerfLoggable()) {
+            Log.perf("Messages adapter - onQueryItemRemoved. Position: " + position);
+        }
     }
 
     @Override
@@ -580,12 +608,20 @@ public class AtlasMessagesAdapter extends RecyclerView.Adapter<AtlasMessagesAdap
         mFooterPosition -= itemCount;
         updateRecipientStatusPosition();
         notifyItemRangeRemoved(positionStart, itemCount);
+
+        if (Log.isPerfLoggable()) {
+            Log.perf("Messages adapter - onQueryItemRangeRemoved. Position start: " + positionStart + " Count: " + itemCount);
+        }
     }
 
     @Override
     public void onQueryItemMoved(RecyclerViewController controller, int fromPosition, int toPosition) {
         updateRecipientStatusPosition();
         notifyItemMoved(fromPosition, toPosition);
+
+        if (Log.isPerfLoggable()) {
+            Log.perf("Conversations adapter - onQueryItemMoved. From: " + fromPosition + " To: " + toPosition);
+        }
     }