+++ layer-atlas/src/main/java/com/layer/atlas/messagetypes/location/LocationSender.java
@@ -72,7 +72,7 @@ public class LocationSender extends AttachmentSender {
         // If the correct Google Play Services are available, connect and return. 
         if (errorCode == ConnectionResult.SUCCESS) {
             GoogleApiCallbacks googleApiCallbacks = new GoogleApiCallbacks();
-            sGoogleApiClient = new GoogleApiClient.Builder(activity)
+            sGoogleApiClient = new GoogleApiClient.Builder(activity.getApplicationContext())
                     .addConnectionCallbacks(googleApiCallbacks)
                     .addOnConnectionFailedListener(googleApiCallbacks)
                     .addApi(LocationServices.API)
@@ -136,7 +136,7 @@ public class LocationSender extends AttachmentSender {
         if (activity == null) return false;
         if (Log.isLoggable(Log.VERBOSE)) Log.v("Sending location");
         if (checkSelfPermission(activity, PERMISSION) != PackageManager.PERMISSION_GRANTED) {
-            requestPermissions(activity, new String[]{PERMISSION}, PERMISSION_REQUEST_CODE);
+            requestPermissions(activity, PERMISSION_REQUEST_CODE, PERMISSION);
             return true;
         }
         return getFreshLocation(new SenderLocationListener(this));
@@ -159,6 +159,10 @@ public class LocationSender extends AttachmentSender {
         @Override
         public void onLocationChanged(Location location) {
             if (Log.isLoggable(Log.VERBOSE)) Log.v("Got fresh location");
+
+            if (Log.isPerfLoggable()) {
+                Log.perf("LocationSender is attempting to send a message");
+            }
             LocationSender sender = mLocationSenderReference.get();
             if (sender == null) return;
             Context context = sender.getContext();