+++ layer-atlas/src/main/java/com/layer/atlas/messagetypes/threepartimage/CameraSender.java
@@ -1,11 +1,18 @@
 package com.layer.atlas.messagetypes.threepartimage;
 
+import android.Manifest;
 import android.app.Activity;
 import android.content.Intent;
+import android.content.pm.PackageManager;
 import android.net.Uri;
+import android.os.Build;
 import android.os.Bundle;
+import android.os.Environment;
 import android.os.Parcelable;
 import android.provider.MediaStore;
+import android.support.annotation.NonNull;
+import android.support.v4.content.FileProvider;
+import android.text.TextUtils;
 
 import com.layer.atlas.R;
 import com.layer.atlas.messagetypes.AttachmentSender;
@@ -22,7 +29,7 @@ import java.util.concurrent.atomic.AtomicReference;
 
 /**
  * CameraSender creates a ThreePartImage from the device's camera.
- *
+ * <p>
  * Note: If your AndroidManifest declares that it uses the CAMERA permission, then CameraSender will
  * require that the CAMERA permission is also granted.  If your AndroidManifest does not declare
  * that it uses the CAMERA permission, then CameraSender will not require the CAMERA permission to
@@ -30,28 +37,40 @@ import java.util.concurrent.atomic.AtomicReference;
  * for details.
  */
 public class CameraSender extends AttachmentSender {
+    private static final String PERMISSION_READ = (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN) ? Manifest.permission.READ_EXTERNAL_STORAGE : null;
+    private static final String PERMISSION_WRITE = (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN) ? Manifest.permission.WRITE_EXTERNAL_STORAGE : null;
+
     public static final int ACTIVITY_REQUEST_CODE = 20;
+    public static final int PERMISSION_REQUEST_CODE = 110;
 
     private WeakReference<Activity> mActivity = new WeakReference<Activity>(null);
 
     private final AtomicReference<String> mPhotoFilePath = new AtomicReference<String>(null);
+    private final String mFileProviderAuthority;
 
-    public CameraSender(int titleResId, Integer iconResId, Activity activity) {
-        this(activity.getString(titleResId), iconResId, activity);
+    public CameraSender(int titleResId, Integer iconResId, Activity activity, @NonNull String fileProviderAuthority) {
+        this(activity.getString(titleResId), iconResId, activity, fileProviderAuthority);
     }
 
-    public CameraSender(String title, Integer iconResId, Activity activity) {
+    public CameraSender(String title, Integer iconResId, Activity activity, @NonNull String fileProviderAuthority) {
         super(title, iconResId);
         mActivity = new WeakReference<Activity>(activity);
+        if (TextUtils.isEmpty(fileProviderAuthority)) {
+            throw new IllegalArgumentException("Empty file provider authority");
+        }
+        mFileProviderAuthority = fileProviderAuthority;
     }
 
     private void startCameraIntent(Activity activity) {
         String fileName = "cameraOutput" + System.currentTimeMillis() + ".jpg";
-        File file = new File(getContext().getExternalFilesDir(android.os.Environment.DIRECTORY_PICTURES), fileName);
+        File file = new File(Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_PICTURES), fileName);
         mPhotoFilePath.set(file.getAbsolutePath());
+
+        final Uri outputUri = FileProvider.getUriForFile(activity, mFileProviderAuthority, file);
+
         Intent cameraIntent = new Intent(android.provider.MediaStore.ACTION_IMAGE_CAPTURE);
-        final Uri outputUri = Uri.fromFile(file);
         cameraIntent.putExtra(MediaStore.EXTRA_OUTPUT, outputUri);
+
         activity.startActivityForResult(cameraIntent, ACTIVITY_REQUEST_CODE);
     }
 
@@ -59,9 +78,33 @@ public class CameraSender extends AttachmentSender {
     public boolean requestSend() {
         Activity activity = mActivity.get();
         if (activity == null) return false;
-        if (Log.isLoggable(Log.VERBOSE)) Log.v("Sending camera image");
-        startCameraIntent(activity);
+
+        if (Log.isLoggable(Log.VERBOSE)) Log.v("Checking permissions");
+
+        if (!hasPermissions(activity, PERMISSION_READ, PERMISSION_WRITE)) {
+            requestPermissions(activity, PERMISSION_REQUEST_CODE, PERMISSION_READ, PERMISSION_WRITE);
+        } else {
+            startCameraIntent(activity);
+        }
+
         return true;
+    }
+
+    @Override
+    public void onRequestPermissionsResult(int requestCode, String[] permissions, int[] grantResults) {
+        if (requestCode != PERMISSION_REQUEST_CODE) return;
+
+        if (grantResults.length != 2) {
+            if (Log.isLoggable(Log.VERBOSE)) Log.v("One or more required  permissions denied");
+            return;
+        }
+
+        if (grantResults[0] == PackageManager.PERMISSION_GRANTED && grantResults[1] == PackageManager.PERMISSION_GRANTED) {
+            if (Log.isLoggable(Log.VERBOSE)) Log.v("Sending camera image");
+            Activity activity = mActivity.get();
+            if (activity == null) return;
+            startCameraIntent(activity);
+        }
     }
 
     @Override
@@ -73,6 +116,9 @@ public class CameraSender extends AttachmentSender {
         }
         if (Log.isLoggable(Log.VERBOSE)) Log.v("Received camera response");
         try {
+            if (Log.isPerfLoggable()) {
+                Log.perf("CameraSender is attempting to send a message");
+            }
             Identity me = getLayerClient().getAuthenticatedUser();
             String myName = me == null ? "" : Util.getDisplayName(me);
             Message message = ThreePartImageUtils.newThreePartImageMessage(activity, getLayerClient(), new File(mPhotoFilePath.get()));